Refactor supabase directory
